Finishing 2fa and jwt

+++ backend/src/controllers/BlockedUserController.js
@@ -1,11 +1,26 @@
 const BlockedUserService = require('../services/BlockedUserService');
 const UserService = require('../services/UserService');
+const SECRET_KEY = process.env.JWT_SECRET_KEY;
 
 class BlockedUserController {
 
 	static async createBlockedUser(request, reply) {
 		const { userId, blockedUserId } = request.body;
+		const authHeader = request.headers.authorization;
 		try {
+			if (!authHeader || !authHeader.startsWith('Bearer '))
+				return reply.code(401).send({ message: 'Unauthorized: No token provided' });
+			const accessToken = authHeader.split(' ')[1];
+			let decoded;
+			try {
+				decoded = request.server.jwt.verify(accessToken, SECRET_KEY);
+			} catch (err) {
+				if (err.message.includes('expired'))
+					return reply.code(401).send({ message: "Access token expired!" });
+				return reply.code(401).send({ message: "Invalid access token" });
+			}
+			if (decoded.userId != userId)
+				return reply.code(403).send({ message: "Token does not belong to this user!" });
 			const newblockId = await BlockedUserService.createBlockedUser({ userId, blockedUserId });
 			reply.code(201).send(newblockId);
 		} catch (err) {
@@ -15,7 +30,21 @@ class BlockedUserController {
 
 	static async getAllUserBlocks(request, reply) {
 		const { userId } = request.params;
+		const authHeader = request.headers.authorization;
 		try {
+			if (!authHeader || !authHeader.startsWith('Bearer '))
+				return reply.code(401).send({ message: 'Unauthorized: No token provided' });
+			const accessToken = authHeader.split(' ')[1];
+			let decoded;
+			try {
+				decoded = request.server.jwt.verify(accessToken, SECRET_KEY);
+			} catch (err) {
+				if (err.message.includes('expired'))
+					return reply.code(401).send({ message: "Access token expired!" });
+				return reply.code(401).send({ message: "Invalid access token" });
+			}
+			if (decoded.userId != userId)
+				return reply.code(403).send({ message: "Token does not belong to this user!" });
 			const user = await UserService.getUserById(userId);
 			if (!user)
 				reply.code(404).send({ message: 'User not found!' });
@@ -50,7 +79,19 @@ class BlockedUserController {
 
 	static async deleteBlock(request, reply) {
 		const { id } = request.params;
+		const authHeader = request.headers.authorization;
 		try {
+			if (!authHeader || !authHeader.startsWith('Bearer '))
+				return reply.code(401).send({ message: 'Unauthorized: No token provided' });
+			const accessToken = authHeader.split(' ')[1];
+			let decoded;
+			try {
+				decoded = request.server.jwt.verify(accessToken, SECRET_KEY);
+			} catch (err) {
+				if (err.message.includes('expired'))
+					return reply.code(401).send({ message: "Access token expired!" });
+				return reply.code(401).send({ message: "Invalid access token" });
+			}
 			const changes = await BlockedUserService.deleteBlock(id);
 			if (changes == 0) reply.code(404).send({ message: 'Block not found!' });
 			else reply.code(200).send({ message: 'Block deleted successfully!' });
